Create Dummy Certificate Details Object
