+++ client/src/assets/placeholder/index.ts
@@ -1,2 +1,2 @@
-// eslint-disable-next-line
 export {default as SIGNUP} from './signup';
+export {default as DUMMY_CERTIFICATE} from './certificate';
